+++ src/TruePBR.h
@@ -25,6 +25,8 @@ public:
 	void PrePass();
 	void PostPostLoad();
 	void DataLoaded();
+	bool TESObjectLAND_SetupMaterial(RE::TESObjectLAND* land);
+	bool BSLightingShader_SetupMaterial(RE::BSLightingShader* shader, RE::BSLightingShaderMaterialBase const* material);
 
 	void SetShaderResouces(ID3D11DeviceContext* a_context);
 	void GenerateShaderPermutations(RE::BSShader* shader);
